Very rudimentary system test works
Created "test" directory to house our first test! The
main entry point is test-runner.sh which can be run
from the "test" directory to run the 1 system test
that is currently in place.

This is all very hacky. A major TODO is to convert this
all to python (or maybe javascript) to make it more
manageable.