Create PeerGroup to make it easier to manage a group of peers.
As peers join, addPeer() abstracts the logic necessary to electing
a new initiator and instructing the initiator to offer data channels.

As peers leave, removePeer() abstracts the logic related to removing
a peer from the group.

Currently, whenever a peer joins or leaves the initiator creates a
new set of data channels to its peers, even if the initiator has
not changed. This can be fixed in the future and FIXME's have been
left behind as a reminder.